requirements.txt
# filepath: d:\mcp_servers\weather_mcp\requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.103.0
uvicorn>=0.24.0
pydantic>=2.3.0
fastmcp>=0.2.0